+++ listings/18/018.rs
@@ -1,14 +1,14 @@
 use std::future::Future;
 
-async fn compute(input: u32) -> u32 {
-    input * 2
+async fn compute(input: u32) -> u64 {
+    u64::from(input) * 2
 }
 
-fn compute_desugared(input: u32) -> impl Future<Output = u32> {
-    async move { input * 2 }
+fn compute_desugared(input: u32) -> impl Future<Output = u64> {
+    async move { u64::from(input) * 2 }
 }
 
-fn accepts_future(_future: impl Future<Output = u32>) {}
+fn accepts_future(_future: impl Future<Output = u64>) {}
 
 fn main() {
     accepts_future(compute(21));
